snapshot-recorder/requirements.txt
python-dotenv>=1.0.1
visier-api-data-in==0.99201.1531
visier-api-data-out==0.99201.1531
visier-api-analytic-model==0.99201.1531
visier-api-core==0.99201.1531
certifi